Change data model to fit apiv3

+++ android/src/test/java/com/thebluealliance/androidclient/datafeed/maps/AddDistrictKeysTest.java
@@ -41,7 +41,7 @@ public class AddDistrictKeysTest {
         assertEquals(districtList.size(), 1);
         assertEquals(districtList.get(0), mDistrict);
 
-        assertEquals(mDistrict.getYear(), mYear);
+        assertEquals(mDistrict.getYear().intValue(), mYear);
         assertEquals(mDistrict.getKey(), DistrictHelper.generateKey(mDistrict.getAbbreviation(), mYear));
     }
 }
